Liste.c
Let M be a matrix of integers M of size NxN, M has the particularity to contain many null values (hollow matrix). These insignificant data (boxes) occupy memory anyway at the level of the matrix. The goal of this project is to use chained lists to represent the elements of the matrix M without representing the null values in order to save space.

Task: You are asked to extract a list from the matrix M

Explanation of the structure: the vertical list represents the list of rows and each element of this list contains a pointer to a horizontal list containing the columns of the current row i with non-zero values.
Example : The structure above is built from the following matrix M of size 3x3:


M :  0    10    15
     12    0    16
     0     0    12
1-Give the structures (records) needed to make this list of lists.
2-Implement a function that displays this list of lists (think of a simple way of representation).
3-Implement a function liberateList (list * l) which allows to delete all the structure created. Use this function at the end of the program.
4-Write a function int getVal(List *l, int i, int j) which allows to return the value of the cell (i,j) of the matrix from the list. Treat the particular cases : i and j are not valid (values higher than M or negative values), the case where the element (i,j) does not exist in the whole list, it means that this element is worth 0 so return 0.
5-Implement a main program which starts by entering the elements of the matrix then create the list of lists from the matrix, display the list, call the getVal function to display the values of some elements according to the user's input. Delete (the list) before leaving the program.